# api/api-spec.yaml
openapi: 3.0.3
info:
  title: TripleThink API
  description: |
    Event-sourced narrative construction system for multi-book fiction series.

    TripleThink enables authors to manage complex storylines with:
    - Epistemic tracking (who knows what, when)
    - Fiction management (false narratives with strict audience constraints)
    - Time-travel queries (entity state at any timestamp)
    - Token-efficient AI integration

    ## Key Concepts

    - **Entities**: Characters, events, objects, locations, systems
    - **Metadata**: Rich author notes and AI guidance (loaded on-demand)
    - **Knowledge States**: What characters believe at specific times
    - **Fictions**: False narratives with target audiences
    - **Epistemic Queries**: The power feature for tracking beliefs
  version: 1.0.0
  contact:
    name: TripleThink Support
  license:
    name: MIT
servers:
  - url: http://localhost:3000
    description: Local development server
...
paths:
...
  /api/metadata:
    get:
      tags: [Metadata]
      summary: List metadata entries
      parameters:
        - name: entity_type
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 100
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        '200':
          description: List of metadata
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Metadata'
                  pagination:
                    $ref: '#/components/schemas/Pagination'

    post:
      tags: [Metadata]
      summary: Create metadata
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Metadata'
      responses:
        '201':
          description: Metadata created
        '400':
          $ref: '#/components/responses/ValidationError'
  /api/metadata/{id}:
    get:
      tags: [Metadata]
      summary: Get metadata by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Metadata found
        '404':
          $ref: '#/components/responses/NotFound'

    put:
      tags: [Metadata]
      summary: Update metadata
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Metadata'
      responses:
        '200':
          description: Metadata updated
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      tags: [Metadata]
      summary: Delete metadata
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Metadata deleted
        '404':
          $ref: '#/components/responses/NotFound'

  # ==========================================
  # EPISTEMIC (POWER FEATURE)
  # ==========================================
...
components:
  schemas:
...
    Metadata:
      type: object
      properties:
        id:
          type: string
        entity_id:
          type: string
        entity_type:
          type: string
        author_notes:
          type: object
          properties:
            creative_intent:
              type: string
            themes:
              type: array
              items:
                type: string
            constraints:
              type: array
              items:
                type: string
        ai_guidance:
          type: object
          properties:
            tone:
              type: string
            scene_rendering:
              type: string
            pov_recommendations:
              type: array
              items:
                type: string
        dev_status:
          type: object
          properties:
            completeness:
              type: string
            todo:
              type: array
              items:
                type: string
            warnings:
              type: array
              items:
                type: string
        version_info:
          type: object
          required: [created, modified]
          properties:
            created:
              type: string
              format: date-time
            modified:
              type: string
              format: date-time
            changelog:
              type: array
              items:
                type: object
        prose_guidance:
          type: object
          properties:
            voice:
              type: string
            tone:
              type: string
            pacing_notes:
              type: string
        consistency_rules:
          type: array
          items:
            type: string
...
    Pagination:
      type: object
      properties:
        limit:
          type: integer
        offset:
          type: integer
        count:
          type: integer
        total:
          type: integer
...
    Error:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string
            details:
              type: object
            suggestion:
              type: string
            timestamp:
              type: string
              format: date-time
  responses:
    ValidationError:
      description: Invalid input
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error:
              code: VALIDATION_ERROR
              message: Missing required fields
              details:
                missing_fields: [id, name]
              suggestion: Check request body for required fields
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error:
              code: NOT_FOUND
              message: 'Entity not found: char-unknown'
              suggestion: Verify the ID is correct
...
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
security:
  - ApiKeyAuth: []